% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_demo.R
\name{board_demo}
\alias{board_demo}
\title{Generate a demo pin board}
\usage{
board_demo()
}
\description{
Generate a \code{\link[pins:board_folder]{pins::board_temp()}} using R's built-in datasets \code{mtcars},
\code{BJsales}, \code{beaver1}, and some of their metadata.
}
\examples{
my_board <- board_demo()
}
