% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinbrowser-package.R
\name{pinbrowser}
\alias{pinbrowser}
\title{Interactively browse pins and their metadata}
\description{
Browse the content of boards created with the pins package. View and search
available pins, view their metadata, and generate code to read pins into your
current project.
}
\keyword{internal}
